requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
cryptography==42.0.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
httpx==0.26.0
